refactor: extract TypeScript types from api.ts into types.ts (#160)
Move all type/interface definitions to a dedicated types.ts module.
api.ts re-exports via `export type *` for backwards compatibility.
